♻️  zb: Make `fdo` a hierarchy
This not only cleans up the `fdo` module by splitting it into multiple
files, it also fixes the issue of the blocking `fdo` proxies getting
exposed under the `crate::fdo` module with commit e7b72ae103 (🔥 zb:
Drop use of macro_rules for fdo proxies).

This also means that blocking iterators are now under `blocking::fdo`
module, as they always should have been.
